+++ src/utils/start/src/index.action.ts
@@ -1,5 +1,5 @@
 /* tslint:disable:max-line-length */
-import {canRunCommand, Format, isMacOS, locateBinaryMacOS, Log} from '@diez/cli-core';
+import {canRunCommand, Format, getPackageManager, isChildProcess, isMacOS, locateBinaryMacOS, Log} from '@diez/cli-core';
 import {Target} from '@diez/engine';
 import {ChildProcess, execSync, fork, spawn} from 'child_process';
 import {readdirSync} from 'fs-extra';
@@ -14,16 +14,9 @@ const guideUrls = {
   [Target.Web]: 'https://diez.org/getting-started/javascript.html',
 };
 
-export = async (_: {}, target: Target, targetRootIn?: string) => {
+export = async (_: {}, target: Target) => {
   if (![Target.Android, Target.Web, Target.Ios].includes(target)) {
     Log.error(`Usage: diez start <${Target.Android}|${Target.Ios}|${Target.Web}>`);
-    process.exit(1);
-    return;
-  }
-
-  // Make sure Yarn is installed.
-  if (!await canRunCommand('yarn --version')) {
-    Log.error('Yarn is required to run the example projects. See https://yarnpkg.org for details.');
     process.exit(1);
     return;
   }
@@ -47,28 +40,29 @@ export = async (_: {}, target: Target, targetRootIn?: string) => {
 
   const diez = require.resolve('diez');
   const root = global.process.cwd();
-  const targetRoot = targetRootIn || resolve(root, '..', 'example-codebases', target);
+  const targetRoot = resolve(root, '..', 'example-codebases', target);
+  const packageManager = await getPackageManager();
 
   Log.comment(`Building Diez project for target ${target}...`);
   let hotProcess!: ChildProcess;
   const guideUrl = guideUrls[target];
   switch (target) {
     case Target.Android:
-      execSync('yarn diez compile -t android', {stdio: 'inherit'});
+      execSync(`${diez} compile -t android`, {stdio: 'inherit'});
       Log.comment('Starting the Diez hot server...');
       hotProcess = fork(diez, ['hot', '-t', 'android'], {stdio: 'inherit'});
       break;
     case Target.Ios:
-      execSync('yarn diez compile -t ios --cocoapods', {stdio: 'inherit'});
+      execSync(`${diez} compile -t ios --cocoapods`, {stdio: 'inherit'});
       Log.comment('Installing CocoaPods dependencies in example codebase...');
       execSync('pod install', {cwd: targetRoot, stdio: 'inherit'});
       Log.comment('Starting the Diez hot server...');
       hotProcess = fork(diez, ['hot', '-t', 'ios'], {stdio: 'inherit'});
       break;
     case Target.Web:
-      execSync('yarn diez compile -t web', {stdio: 'inherit'});
+      execSync(`${diez} compile -t web`, {stdio: 'inherit'});
       Log.comment('Installing Node dependencies in example codebase...');
-      execSync('yarn', {cwd: targetRoot, stdio: 'inherit'});
+      await packageManager.installAllDependencies({cwd: targetRoot, stdio: 'inherit'});
       Log.comment('Starting the Diez hot server...');
       hotProcess = fork(diez, ['hot', '-t', 'web'], {stdio: 'inherit'});
       break;
@@ -109,20 +103,23 @@ To learn more, follow along with the guide at:
           return;
         }
       case Target.Web:
-        return spawn(
-          'yarn',
-          ['start'],
-          {cwd: targetRoot, stdio: 'inherit', shell: true},
-        );
+        return spawn(packageManager.binary, ['start'], {cwd: targetRoot, stdio: 'inherit', shell: true});
     }
   };
+
+  let appProcess: void | ChildProcess | Buffer;
 
   const handleBuilt = (message: string) => {
     if (message === 'built') {
       hotProcess.removeListener('message', handleBuilt);
-      runApp();
-
+      appProcess = runApp();
     }
   };
+
   hotProcess.on('message', handleBuilt);
+  hotProcess.on('exit', () => {
+    if (isChildProcess(appProcess)) {
+      appProcess.kill();
+    }
+  });
 };