+++ src/utils/start/test/start.test.ts
@@ -1,5 +1,6 @@
-import {assignMock, cleanupMockCommandData, diezRun, mockCanRunCommand, mockCliCoreFactory} from '@diez/test-utils';
-jest.doMock('@diez/cli-core', mockCliCoreFactory);
+import {assignMock, cleanupMockCommandData, diezRun, mockCanRunCommand, mockCliCoreFactory, mockPackageManagerInstance} from '@diez/test-utils';
+const cliCore = mockCliCoreFactory();
+jest.doMock('@diez/cli-core', () => cliCore);
 
 const mockFork = jest.fn();
 const mockSpawn = jest.fn();
@@ -23,8 +24,8 @@ beforeEach(() => {
   mockFork.mockReset();
   mockFork.mockReturnValue(mockProcess);
   mockSpawn.mockReset();
-  mockExecSync.mockReset();
-  mockExecSync.mockReturnValue('');
+  mockExecSync.mockClear();
+  mockPackageManagerInstance.exec.mockClear();
 });
 
 afterEach(() => {
@@ -37,10 +38,24 @@ describe('diez start command', () => {
     expect(process.exit).toHaveBeenCalledWith(1);
   });
 
-  test('crashes if no yarn', async () => {
-    mockCanRunCommand.mockResolvedValue(false);
+  test('kills app process if start exits', async () => {
+    const appProcess = {
+      killed: false,
+      kill () {
+        Object.assign(this, {killed: true});
+      },
+    };
+
+    mockSpawn.mockReturnValueOnce(appProcess);
+
     await diezRun('start web');
-    expect(process.exit).toHaveBeenCalledWith(1);
+    expect(mockFork).toHaveBeenNthCalledWith(
+      1, expect.stringContaining('diez/bin/diez'), ['hot', '-t', 'web'], expect.anything());
+
+    mockProcess.emit('message', 'built');
+    mockProcess.emit('exit');
+    expect(appProcess.killed).toBe(true);
+    expect(process.exit).toHaveBeenCalled();
   });
 
   test('Android golden path', async () => {
@@ -59,19 +74,19 @@ describe('diez start command', () => {
   });
 
   test('iOS - bad CocoaPods', async () => {
-    mockExecSync.mockReturnValue('1.6.9');
+    mockExecSync.mockReturnValueOnce('1.6.9');
     await diezRun('start ios');
     expect(process.exit).toHaveBeenCalledWith(1);
   });
 
   test('iOS golden path', async () => {
-    mockExecSync.mockReturnValue('1.7.0');
+    mockExecSync.mockReturnValueOnce('1.7.0');
     await diezRun('start ios');
     mockProcess.emit('message', 'built');
     expect(mockExecSync).toHaveBeenNthCalledWith(1, 'pod --version');
     expect(mockExecSync).toHaveBeenNthCalledWith(
-      2, expect.stringContaining('diez compile -t ios --cocoapods'), expect.anything());
-    expect(mockExecSync).toHaveBeenNthCalledWith(3, 'pod install', expect.anything());
+      7, expect.stringContaining('diez compile -t ios --cocoapods'), expect.anything());
+    expect(mockExecSync).toHaveBeenLastCalledWith('pod install', expect.anything());
     expect(mockFork).toHaveBeenCalledWith(
       expect.stringContaining('diez/bin/diez'), ['hot', '-t', 'ios'], expect.anything());
   });