+++ src/utils/test-utils/src/mocks/@diez/cli-core.ts
@@ -32,6 +32,55 @@ export const cleanupMockCommandData = () => {
 };
 
 /**
+ * A mock singleton for `shouldUseYarn` in `@diez/cli-core`.
+ */
+export const mockShouldUseYarn = jest.fn();
+
+/**
+ * A mock singleton for `packageManager` in `@diez/cli-core`.
+ */
+export const mockPackageManagerInstance = {
+  binary: 'yarn',
+  exec: jest.fn(() => new Promise((resolve) => resolve({}))),
+  installAllDependencies: jest.fn(() => new Promise((resolve) => resolve({}))),
+};
+
+/**
+ * A mock singleton for `Log.error` in `@diez/cli-core`.
+ */
+export const mockLogError = jest.fn();
+
+/**
+ * A mock singleton for `Log.code` in `@diez/cli-core`.
+ */
+export const mockLogCode = jest.fn();
+
+/**
+ * A mock singleton for `Log.info` in `@diez/cli-core`.
+ */
+export const mockLogInfo = jest.fn();
+
+/**
+ * A mock singleton for `Log.comment` in `@diez/cli-core`.
+ */
+export const mockLogComment = jest.fn();
+
+/**
+ * A mock singleton for `Log.infoTitle` in `@diez/cli-core`.
+ */
+export const mockLogInfoTitle = jest.fn();
+
+/**
+ * A mock singleton for `Log.warning` in `@diez/cli-core`.
+ */
+export const mockLogWarning = jest.fn();
+
+/**
+ * A mock singleton for `Log.warningOnce` in `@diez/cli-core`.
+ */
+export const mockLogWarningOnce = jest.fn();
+
+/**
  * Module mock factory.
  */
 export const mockCliCoreFactory = () => ({
@@ -41,7 +90,20 @@ export const mockCliCoreFactory = () => ({
   socketTrap: mockSocketTrap,
   canRunCommand: mockCanRunCommand,
   locateBinaryMacOS: mockLocateBinaryMacOS,
+  shouldUseYarn: mockShouldUseYarn,
+  Log: {
+    error: mockLogError,
+    code: mockLogCode,
+    info: mockLogInfo,
+    comment: mockLogComment,
+    infoTitle: mockLogInfoTitle,
+    warning: mockLogWarning,
+    warningOnce: mockLogWarningOnce,
+  },
   findOpenPort () {
     return Promise.resolve(9001);
+  },
+  getPackageManager () {
+    return Promise.resolve(mockPackageManagerInstance);
   },
 });